pyproject.toml
[project]
name = "turboagent-minimal-demo"
version = "0.1.0"
description = "Minimal Pydantic AI + SurrealDB + TurboAgents demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.1.0",
  "pydantic-ai-slim[openai]>=0.8.0",
  "surrealdb>=1.0.0",
  "turboagents[rag]>=0.1.0a2",
]

[tool.uv]
package = false
